move percentile_counter
