+++ ydb/core/persqueue/pqtablet/partition/user_info.h
@@ -4,7 +4,7 @@
 
 #include <ydb/core/persqueue/working_time_counter.h>
 #include "subscriber.h"
-#include <ydb/core/persqueue/percentile_counter.h>
+#include <ydb/core/persqueue/public/percentile_counter.h>
 #include "quota_tracker.h"
 #include <ydb/core/persqueue/public/constants.h>
 #include <ydb/core/persqueue/pqtablet/metering_sink.h>
