+++ ydb/core/persqueue/public/ya.make
@@ -1,6 +1,7 @@
 LIBRARY()
 
 SRCS(
+    percentile_counter.cpp
     pq_database.cpp
     pq_rl_helpers.cpp
     utils.cpp
