+++ ydb/core/persqueue/ya.make
@@ -5,7 +5,6 @@ SRCS(
     heartbeat.cpp
     key.cpp
     list_all_topics_actor.cpp
-    percentile_counter.cpp
     write_meta.cpp
     dread_cache_service/caching_service.cpp
     write_id.cpp
