+++ ydb/services/deprecated/persqueue_v0/grpc_pq_actor.h
@@ -20,7 +20,7 @@
 #include <ydb/core/persqueue/events/global.h>
 #include <ydb/core/persqueue/writer/partition_chooser.h>
 #include <ydb/core/persqueue/writer/writer.h>
-#include <ydb/core/persqueue/percentile_counter.h>
+#include <ydb/core/persqueue/public/percentile_counter.h>
 #include <ydb/core/base/appdata.h>
 #include <ydb/core/base/tablet_pipe.h>
 #include <ydb/core/tx/tx_proxy/proxy.h>
